refactor: aop 코드 경로 변경 (#126)
notification에서 notification.aop로 변경
